Fix index type (ssize_t)

+++ lib/collections.c
@@ -111,7 +111,8 @@ void queue_free(Queue* queue)
     while (!queue_is_empty(queue))
     {
         void* data = queue_dequeue(queue);
-        if (queue->free_function) {
+        if (queue->free_function)
+        {
             queue->free_function(data);
         }
     }
@@ -307,9 +308,9 @@ Dictionary* dictionary_init(void)
     return dictionary;
 }
 
-static uint64_t dictionary_lookup(Dictionary* dictionary, const char* key)
+static ssize_t dictionary_lookup(Dictionary* dictionary, const char* key)
 {
-    uint64_t index = hash_key(key) % dictionary->max_length;
+    ssize_t index = hash_key(key) % dictionary->max_length;
     while (dictionary->key_value_pairs[index].key != NULL)
     {
         if (strcmp(dictionary->key_value_pairs[index].key, key) == 0)
@@ -371,7 +372,7 @@ static void dictionary_rebuild(Dictionary* dictionary)
 
 void dictionary_set(Dictionary* dictionary, const char* key, void* value)
 {
-    uint64_t index = dictionary_lookup(dictionary, key);
+    ssize_t index = dictionary_lookup(dictionary, key);
     // Key does not exist, add new KeyValuePair
     if (index == KEY_NOT_FOUND)
     {
@@ -404,7 +405,7 @@ void dictionary_set(Dictionary* dictionary, const char* key, void* value)
 
 void* dictionary_get(Dictionary* dictionary, const char* key)
 {
-    uint64_t index = dictionary_lookup(dictionary, key);
+    ssize_t index = dictionary_lookup(dictionary, key);
     // Key successfully found, return value
     if (index != KEY_NOT_FOUND)
     {
